# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Panel26 Admin API",
        "description": "Panel26 Admin Panel REST API - Bearer token and API Key authentication supported. All endpoints accept ?lang=tr|en query parameter for localized responses.",
        "contact": {
            "email": "admin@panel26.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API V1 Server"
        }
    ],
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login with email and password",
                "operationId": "a7997e7c1e3bfed64d4e56de82415931",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@panel26.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "iPhone 15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "429": {
                        "description": "Too many attempts"
                    }
                }
            }
        },
        "/api/v1/auth/2fa/verify": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify 2FA code after login",
                "operationId": "9603911543c0af235000cf8c6a3dea2a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "temp_token",
                                    "code"
                                ],
                                "properties": {
                                    "temp_token": {
                                        "type": "string"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "2FA verified"
                    },
                    "401": {
                        "description": "Invalid code"
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout and revoke current token",
                "operationId": "f6b8957b4c0b25636613a9f0512a8ba9",
                "responses": {
                    "200": {
                        "description": "Logged out"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get current authenticated user",
                "operationId": "37d33f0281f67d31d05b4c45eea0143b",
                "responses": {
                    "200": {
                        "description": "Current user data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/tokens": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "List user tokens",
                "operationId": "05bc8b5814457db2031ac48b61cb52dc",
                "responses": {
                    "200": {
                        "description": "Token list"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/tokens/{id}": {
            "delete": {
                "tags": [
                    "Auth"
                ],
                "summary": "Revoke a token",
                "operationId": "042c7380c777a77a7d0d80febc942829",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token revoked"
                    },
                    "404": {
                        "description": "Token not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/account": {
            "delete": {
                "tags": [
                    "Auth"
                ],
                "summary": "Delete current user account (GDPR compliant)",
                "operationId": "667ec795ae736092cd302803313b5465",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Account deleted"
                    },
                    "401": {
                        "description": "Invalid password"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/account/delete-request": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Request account deletion (sends confirmation email)",
                "operationId": "962e9babb93167f465a390f914055c06",
                "responses": {
                    "200": {
                        "description": "Confirmation email sent"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/roles": {
            "get": {
                "tags": [
                    "General"
                ],
                "summary": "List roles",
                "operationId": "30e8b9c4728f8bbb1a01ed1eaf5c6809",
                "responses": {
                    "200": {
                        "description": "Role list"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/activity-logs": {
            "get": {
                "tags": [
                    "General"
                ],
                "summary": "List activity logs",
                "operationId": "c0262c1cbe6a251a67f5d34a60b49246",
                "parameters": [
                    {
                        "name": "module",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Log list"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Sanctum Bearer Token",
                "bearerFormat": "token",
                "scheme": "bearer"
            },
            "apiKeyAuth": {
                "type": "apiKey",
                "description": "API Key authentication",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}